+++ ja2lib/Tactical/TacticalSave.h
@@ -16,7 +16,7 @@ enum {
 };
 
 // Add
-BOOLEAN AddMapModification(INT16 sMapX, INT16 sMapY, INT8 bMapZ);
+BOOLEAN AddMapModification(u8 sMapX, u8 sMapY, i8 bMapZ);
 
 // Load the Map modifications from the saved game file
 BOOLEAN LoadMapTempFilesFromSavedGameFile(HWFILE hFile);
@@ -25,10 +25,10 @@ BOOLEAN LoadMapTempFilesFromSavedGameFile(HWFILE hFile);
 BOOLEAN SaveMapTempFilesToSavedGameFile(HWFILE hFile);
 
 // delete temp file
-BOOLEAN DeleteTempItemMapFile(INT16 sMapX, INT16 sMapY, INT8 bMapZ);
+BOOLEAN DeleteTempItemMapFile(u8 sMapX, u8 sMapY, i8 bMapZ);
 
 // Retrieves the number of items in the sectors temp item file
-BOOLEAN GetNumberOfWorldItemsFromTempItemFile(INT16 sMapX, INT16 sMapY, INT8 bMapZ,
+BOOLEAN GetNumberOfWorldItemsFromTempItemFile(u8 sMapX, u8 sMapY, i8 bMapZ,
                                               UINT32 *puiNumberOfItems, BOOLEAN fIfEmptyCreate);
 
 // Saves the Current Sectors, ( world Items, rotting corpses, ... )  to the temporary file used to
@@ -57,7 +57,7 @@ BOOLEAN AddWorldItemsToUnLoadedSector(u8 sMapX, u8 sMapY, INT8 bMapZ, INT16 sGri
 BOOLEAN InitTacticalSave(BOOLEAN fCreateTempDir);
 
 // Gets the number of ACTIVE ( Not the TOTAL number ) of World Items from the sectors temp file
-BOOLEAN GetNumberOfActiveWorldItemsFromTempFile(INT16 sMapX, INT16 sMapY, INT8 bMapZ,
+BOOLEAN GetNumberOfActiveWorldItemsFromTempFile(u8 sMapX, u8 sMapY, i8 bMapZ,
                                                 UINT32 *pNumberOfData);
 
 // Call this function to set the new sector a NPC will travel to
@@ -65,7 +65,7 @@ void ChangeNpcToDifferentSector(UINT8 ubNpcId, u8 sSectorX, u8 sSectorY, INT8 bS
 
 // Adds a rotting corpse definition to the end of a sectors rotting corpse temp file
 BOOLEAN AddRottingCorpseToUnloadedSectorsRottingCorpseFile(
-    INT16 sMapX, INT16 sMapY, INT8 bMapZ, ROTTING_CORPSE_DEFINITION *pRottingCorpseDef);
+    u8 sMapX, u8 sMapY, i8 bMapZ, ROTTING_CORPSE_DEFINITION *pRottingCorpseDef);
 
 // Flags used for the AddDeadSoldierToUnLoadedSector() function
 #define ADD_DEAD_SOLDIER_USE_GRIDNO \
@@ -109,11 +109,11 @@ void InitExitGameDialogBecauseFileHackDetected();
 
 void HandleAllReachAbleItemsInTheSector(u8 sSectorX, u8 sSectorY, INT8 bSectorZ);
 
-void GetMapTempFileName(UINT32 uiType, STR pMapName, INT16 sMapX, INT16 sMapY, INT8 bMapZ);
+void GetMapTempFileName(UINT32 uiType, STR pMapName, u8 sMapX, u8 sMapY, i8 bMapZ);
 
 UINT32 GetNumberOfVisibleWorldItemsFromSectorStructureForSector(INT16 sMapX, INT16 sMapY,
                                                                 INT8 bMapZ);
-void SetNumberOfVisibleWorldItemsInSectorStructureForSector(INT16 sMapX, INT16 sMapY, INT8 bMapZ,
+void SetNumberOfVisibleWorldItemsInSectorStructureForSector(u8 sMapX, u8 sMapY, i8 bMapZ,
                                                             UINT32 uiNumberOfItems);
 
 #define NEW_ROTATION_ARRAY_SIZE 49
